requirements.txt
pymodbus==3.6.9
paho-mqtt==1.6.1
influxdb-client==1.43.0
PyYAML==6.0.2
python-dotenv==1.0.0
